+++ examples/svelte-vite-plugin/src/lib/Prose.svelte
@@ -1,11 +1,15 @@
 <script lang="ts">
   import { marked } from "marked";
+  marked.use({
+    mangle: false,
+    headerIds: false
+  });
 </script>
 
 <div
   class="text-base prose prose-purple md:prose-orange dark:prose-invert max-w-full text-left mt-10 p-3 md:p-5 shadow-xl border rounded"
 >
-  {@html marked(`# Svelte Scoped Uno with Vite + Svelte
+  {@html marked.parse(`# Svelte Scoped Uno with Vite + Svelte
   This project has the following **example usages**:
   
 - dark mode