+++ examples/sveltekit-vite-plugin/src/routes/Prose.svelte
@@ -7,7 +7,7 @@
 </script>
 
 <div
-  class="text-base prose prose-purple md:prose-orange  dark:prose-invert max-w-full text-left mt-10 p-3 md:p-5 shadow-lg border rounded"
+  class="text-base prose prose-purple md:prose-orange dark:prose-invert max-w-full text-left mt-10 p-3 md:p-5 shadow-lg border rounded"
 >
   {@html marked.parse(`# Svelte Scoped Uno with SvelteKit
   This project has the following **example usages**: