+++ packages/svelte-preprocess-unocss/CHANGELOG.md
@@ -1,5 +1,13 @@
 # svelte-preprocess-unocss
 
+## 0.1.4
+
+### Patch Changes
+
+- better module resolution
+- Updated dependencies
+  - svelte-scoped-uno@0.1.4
+
 ## 0.1.3
 
 ### Patch Changes